fixing delete account issue

+++ controllers/c_delete_account.php
@@ -2,6 +2,10 @@
 
 require_once(PATH_MODELS . 'UserDAO.php');
 $user = new UserDAO(true);
-$user->deleteUser($_SESSION['user_id']);
-$_SESSION = array();
-header("Location: index.php?page=login");
+if (isset($_GET['verif'])) {
+    if (($_GET['verif'] == 214428544845492) && isset($_SESSION['user_id'])) {
+        $user->deleteUser($_SESSION['user_id']);
+        $_SESSION = array();
+    }
+}
+header("Location: index.php?page=home");
